Move apiClient.ts to folder and hide image upload
